pyproject.toml
[project]
name = "fast-flights-mcp"
version = "0.1.0"
description = "MCP server wrapping fast-flights API"
requires-python = ">=3.11"
dependencies = [
    "fast-flights>=2.2",
    "fastmcp",
    "mcp[cli]",
    "requests",
    "py-moneyed",
    "Babel",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]
dev = [
    "pre-commit",
    "black",
    "ruff",
] 

[project.scripts]
fast-flights-mcp = "fast_flights_mcp.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-vv"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fast_flights_mcp"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["fast_flights_mcp", "src"]
